refactor presets location
