+++ lib/config/presets/default/index.js
@@ -158,6 +158,15 @@ module.exports = {
       },
     ],
   },
+  group: {
+    description: 'Group {{arg2}} packages',
+    packageRules: [
+      {
+        extends: ['{{arg1}}'],
+        groupName: '{{arg2}}',
+      },
+    ],
+  },
   base: {
     description: 'Default base configuration for repositories',
     extends: [
