+++ lib/config/presets/default/readme.md
@@ -1,0 +1,3 @@
+# default presets
+
+This directory contains default presets for Renovate. Any preset defined here can be referenced without a prefix, e.g. `"extends": ["base"]` and not `"extends": ["default/base"]`.
