+++ lib/config/presets/index.js
@@ -1,7 +1,7 @@
-const presetDefaults = require('./presets/default');
-const presetPackages = require('./presets/packages');
-const configParser = require('./index');
-const massage = require('./massage');
+const presetDefaults = require('./default');
+const presetPackages = require('./packages');
+const configParser = require('../index');
+const massage = require('../massage');
 
 module.exports = {
   resolveConfigPresets,
