+++ lib/config/presets/packages/readme.md
@@ -1,0 +1,5 @@
+# packages presets
+
+This directory contains presets for defining lists of packages.
+
+Preset configs here should include only the following fields from `packageRules`: `packageNames`, `packagePatterns`, `excludePackageNames` and `excludePackagePatterns`. Essentially this means they are partial package rules which exist to be extended elsewhere.
