+++ lib/workers/repository/onboarding.js
@@ -135,7 +135,7 @@ Please make any fixes in _this branch_.
   if (description.length) {
     let configDesc = `---\n\n## Configuration Summary\n\nBased on the currently configured presets, you should expect to see the following behaviour:\n<ul>\n`;
     configDesc +=
-      '  <li>Start dependency updates once this Configure Renovate PR is closed</li>\n';
+      '  <li>Start dependency updates once this Configure Renovate PR is merged or closed</li>\n';
     description.forEach(desc => {
       configDesc += `  <li>${desc}</li>\n`;
     });
